pyproject.toml
[tool.poetry]
name = "api-users"
version = "0.1.0"
description = ""
authors = ["Ramiro Tepehua Cortes <rtcface@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
